TST: Refactoring of test_partition
