+++ src/porepy/applications/test_utils/reference_dense_arrays.py
@@ -2036,3 +2036,170 @@ _dir_inds = np.array(
     ]
 )
 test_fvutils = {"test_bound_exclusion": {"neu_inds": _neu_inds, "dir_inds": _dir_inds}}
+
+# test_partition.py
+test_partition = {"test_cartesian_grids": {}}
+test_partition["test_cartesian_grids"]["4_10"] = {
+    "p_known": np.array(
+            [
+                [0, 0, 1, 1],
+                [0, 0, 1, 1],
+                [0, 0, 1, 1],
+                [2, 2, 3, 3],
+                [2, 2, 3, 3],
+                [2, 2, 3, 3],
+                [4, 4, 5, 5],
+                [4, 4, 5, 5],
+                [4, 4, 5, 5],
+                [4, 4, 5, 5],
+            ]            
+        ).ravel("C")
+}
+test_partition["test_cartesian_grids"]["4_4_4"] = {
+    "p_known": np.array(
+            [
+                [0, 0, 1, 1],
+                [0, 0, 1, 1],
+                [2, 2, 3, 3],
+                [2, 2, 3, 3],
+                [0, 0, 1, 1],
+                [0, 0, 1, 1],
+                [2, 2, 3, 3],
+                [2, 2, 3, 3],
+                [4, 4, 5, 5],
+                [4, 4, 5, 5],
+                [6, 6, 7, 7],
+                [6, 6, 7, 7],
+                [4, 4, 5, 5],
+                [4, 4, 5, 5],
+                [6, 6, 7, 7],
+                [6, 6, 7, 7],
+            ]
+        ).ravel("C")
+}
+test_partition["test_cartesian_grids"]["6_5_4"] = {
+    "p_known": np.array(
+            [
+                0,
+                0,
+                1,
+                1,
+                2,
+                2,
+                0,
+                0,
+                1,
+                1,
+                2,
+                2,
+                3,
+                3,
+                4,
+                4,
+                5,
+                5,
+                3,
+                3,
+                4,
+                4,
+                5,
+                5,
+                3,
+                3,
+                4,
+                4,
+                5,
+                5,
+                0,
+                0,
+                1,
+                1,
+                2,
+                2,
+                0,
+                0,
+                1,
+                1,
+                2,
+                2,
+                3,
+                3,
+                4,
+                4,
+                5,
+                5,
+                3,
+                3,
+                4,
+                4,
+                5,
+                5,
+                3,
+                3,
+                4,
+                4,
+                5,
+                5,
+                6,
+                6,
+                7,
+                7,
+                8,
+                8,
+                6,
+                6,
+                7,
+                7,
+                8,
+                8,
+                9,
+                9,
+                10,
+                10,
+                11,
+                11,
+                9,
+                9,
+                10,
+                10,
+                11,
+                11,
+                9,
+                9,
+                10,
+                10,
+                11,
+                11,
+                6,
+                6,
+                7,
+                7,
+                8,
+                8,
+                6,
+                6,
+                7,
+                7,
+                8,
+                8,
+                9,
+                9,
+                10,
+                10,
+                11,
+                11,
+                9,
+                9,
+                10,
+                10,
+                11,
+                11,
+                9,
+                9,
+                10,
+                10,
+                11,
+                11,
+            ]
+        )
+}
